requirements.txt
covid==2.1.6
pyTelegramBotAPI==3.6.7
country-converter==0.6.7
emoji-country-flag==1.2.1